refactor(llm): wire openai adapter to the codec
Rewire the OpenAI adapter into a thin transport shell over the
`openai_responses` codec from the previous commit. All wire translation
now lives in the codec; the adapter owns only auth (bearer + org/project
headers), base URL, the streaming byte loop, and the codex route config.

- Codex mode splits along the codec seam: encode-side param omission
  rides on `CodecParams::openai_codex`; the transport-side half (blocking
  requests served via streaming) is `RouteConfig.force_streaming`,
  resolved once per call. No provider-name branching — OpenAI's only
  route split is codex.
- Resolve file-backed image attachments via the shared
  `attachments::resolve` (images only: audio/documents render as text
  placeholders in the codec without I/O) before the sync codec encode.
- The streaming byte loop moves to the adapter (same StreamLoop shape as
  the anthropic shell): read `\n\n` blocks, frame with parse_sse_message,
  feed `RawEvent`s to the codec's decoder. The decoder resolves the event
  type from the SSE event line or the JSON `type` field, and `finish()`
  synthesizes nothing — `response.completed`/`incomplete` finish the
  stream, matching the old EOF behavior.
- Migrate the ~54 openai unit tests into the codec submodules they now
  cover: body-shape and provider-options-merge tests assert on
  `encode(ctx).body` (already body-only, no dual-assert splitting
  needed); translate_input/parse_output round-trips go sync; the SSE
  state-machine tests drive `on_event(RawEvent)`. Adapter builders,
  count-tokens HTTP mocks (incl. the operation-label log capture), and
  the codex stream-error propagation test stay in the shell. One new
  test pins the count-tokens endpoint + filtered body on the codec.

Behavior-preserving: the 33 openai_responses wire snapshots (codex,
dual-id, opaque items, attachment drop, streaming) and the full fabro-llm
suite (516 tests) stay green; clippy -D warnings and fmt clean.
